Split APIError.swift functionality into serialization, validation files.
